refactor: Validation 도메인 내용 제거
